refactor(benchmarks): isolate bitmap conformance
